Update for new slimmer template updates
The template is changing to auto-center and provide layout to pages.
This makes these pages ready for that change.

Updated the related element to actually match the markup from the static
version. It had a couple of enhancements over the version that was here
like having the aria-labelled by actually pointing to the right
headings.

It also adds a main element to the page which is used for the layout in
static.