Fix print styles for postal address
When printing, each part of the address was appearing on a single line,
and the first line appeared above it as a header. Anyone printing these
pages and then transcribing the address would likely send things to the
wrong place.

* Don’t use styled spans, use divs (empty divs will have a zero height,
whereas spans with line breaks would create new lines for empty parts)
* Don’t use a heading for the title