Make .phone-detail h4 styling no longer specific to .phone-detail
This is necessary to re-use the same styles for a webchat block with
similarly styled `opening times`. It doesn't seem like this should be
specific to `.phone-detail`, so the path of least resistance is to
move it out. `margin-top` added to make it look correct even when it
doesn't directly follow `.call-charges`.