Fix intermittent expiry test failure
This test would sometimes fail, returning 899 rather than 900 for expiry seconds. This is caused by the `GdsApi::Response.expires_in` method with dynamically calculates this value based on when you call it. It's used in the Contacts controller to set the expiry headers. It would 899 if the request takes longer than a second to process. We stub the value to ensure it always matches the expectation.