Make sure fixtures are always valid
The fixtures currently aren't valid according to the frontend schema
for contacts. This commit adds validation and auto-updating by using
the govuk_schemas gem. If we ever add a required attribute, the random
example generator will generate that value for us. Conversely when we
remove an attribute, we'll be notified by the schema tests failing that
we need to update these fixtures.